Coco's Hangout v2.0
This is the second version for this website.
Using the original files as a template I was able to make some changes that just made so much more appealing, I still need to work on the code that handles creating channels and private portals but that will be a while down the road. So much work, so little time.